# openapi/swagger.yaml
openapi: 3.0.0
info:
  title: WeHOME
  version: 1.0.0
  description: Solve Accommodation Problems
  termsOfService: http://swagger.io/terms/
  contact:
    email: 19020024@vnu.edu.vn
...
servers:
  - url: http://wehome-env.eba-ujwahtza.ap-southeast-1.elasticbeanstalk.com/
  - url: http://localhost:8889
  - url: http://113.167.196.55:8889
...
paths:
...
  /api/user/{userId}:
    get:
      tags:
        - user
      summary: Get user by id
      operationId: user.getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid userId supplied
          content: { }
        '404':
          description: User not found
          content: { }
        '500':
          description: System error
    put:
      tags:
        - user
      summary: Update user
      operationId: user.updateUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '401':
          description: Access token is missing or invalid
        '200':
          description: OK
          content: { }
        '400':
          description: Invalid userId supplied
          content: { }
        '404':
          description: User not found
          content: { }
        '500':
          description: System error
    delete:
      tags:
        - user
      summary: Delete user
      operationId: user.deleteUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        '401':
          description: Access token is missing or invalid
        '200':
          description: OK
          content: { }
        '400':
          description: Invalid userId supplied
          content: { }
        '404':
          description: User not found
          content: { }
        '500':
          description: System error
...
  /api/report/create:
    post:
      tags:
        - report
      security:
        - bearerAuth: [ ]
      summary: Create report
      operationId: report.createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: number
                description:
                  type: string
      responses:
        '401':
          description: Access token is missing or invalid
        default:
          description: OK
  /api/feedback/create:
    post:
      tags:
        - feedback
      security:
        - bearerAuth: [ ]
      summary: Create feedback
      operationId: feedback.createFeedback
      requestBody:
        description: Create feedback object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Feedback'
      responses:
        '401':
          description: Access token is missing or invalid
        '500':
          description: System error
        default:
          description: OK
  /api/facility/create:
    post:
      tags:
        - facility
      security:
        - bearerAuth: [ ]
      summary: Admin create some new facilities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                facilities:
                  type: array
                  items:
                    $ref: '#/components/schemas/Facility'
      responses:
        '401':
          description: Access token is missing or invalid
        '500':
          description: System error
        default:
          description: OK
  /api/facility/delete:
    post:
      tags:
        - facility
      security:
        - bearerAuth: [ ]
      summary: Admin delete some facilities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                facilities:
                  type: array
                  items:
                    type: string
      responses:
        '401':
          description: Access token is missing or invalid
        '500':
          description: System error
        default:
          description: OK
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        userId:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        username:
          type: string
        password:
          type: string
      example:
        user_id: 10013
        name: Trong
        phone: '19001000'
        email: abc123@gmail.com
        username: tieuphonghy01
        password: $2a$08$IeZeErqqW9ZitijjJyFi9O7LxqUt0TEliiHAsXLhWozDIg58wyIjG
        role: client
...
    Feedback:
      type: object
      properties:
        roomId:
          type: number
        clientId:
          type: number
        feedback:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
...
    Facility:
      type: object
      properties:
        facilityId:
          type: number
        facility:
          type: string
        description:
          type: string
          nullable: true
      example:
        facility_id: 5
        facility: Dịch vụ dọn phòng
        description: Dọn phòng vào mỗi buổi sáng